- 모달 밖 영역과 모달은 형제여야한다
