+++ diary/src/pages/Components/MonthModal/MonthEditModal.jsx
@@ -3,7 +3,7 @@ import React, { useState } from 'react'
 import { useDispatch, useSelector } from 'react-redux';
 import Button from '../Button';
 
-const MonthEditModal = ({todo, dayInfo, itemVisible, handleItemModalClose}) => {
+const MonthEditModal = ({ref, modalOutSideClick, todo, dayInfo, itemVisible, handleItemModalClose}) => {
 
     const dispatch = useDispatch();
     const [edited, setEdited] =useState(false)
@@ -23,7 +23,7 @@ const MonthEditModal = ({todo, dayInfo, itemVisible, handleItemModalClose}) => {
       console.log(todo)
   return (
     (itemVisible && dayInfo.locdate ?
-      <div className='z-0 absolute inset-x-auto w-96 h-fit bg-white text-right select-none rounded drop-shadow-2xl'>
+      <div className='z-0 absolute inset-x-auto w-96 h-fit bg-white text-right select-none rounded drop-shadow-2xl' ref={ref} onClick={(e)=>modalOutSideClick(e)}>
         <div className='m-3'>
 
             <Button onClick={()=>setEdited(true)} content="수정" />
